Move BIDS and FSL-related functions to the interfaces module.
